requirements.txt
Django==1.7.2
django-extensions==1.4.9
MySQL-python==1.2.5
boto==2.35.0
django-boto==0.3.8
django-storages==1.1.8
django-pipeline==1.4.2
pillow==2.7.0
djangorestframework==3.0.5
